fix: estandarización de tipos en repositorios Spring Data JPA
 Cambios críticos:
- CrudRepository<Entity, Integer>  CrudRepository<Entity, Long>
- Consistencia total en tipos de ID across toda la capa de persistencia
- Eliminación de incompatibilidades entre entidades y repositorios

 Beneficios:
- Prevención de ClassCastException en operaciones CRUD
- Alineación perfecta con auto-generated Long IDs en PostgreSQL
- Mejor performance en queries con índices bigint
- Código más mantenible y type-safe

+++ src/main/java/com/example/demo/repository/CitaOdontologicaRepository.java
@@ -6,7 +6,7 @@ import com.example.demo.entity.CitaOdontologica;
 import com.example.demo.entity.Odontologo;
 import com.example.demo.entity.Paciente;
 
-public interface CitaOdontologicaRepository extends CrudRepository<CitaOdontologica, Integer> {
+public interface CitaOdontologicaRepository extends CrudRepository<CitaOdontologica, Long> {
 	Iterable<CitaOdontologica> findByOdontologo(Odontologo odontologo);
 
 	Iterable<CitaOdontologica> findByPaciente(Paciente paciente);
