+++ src/main/java/com/example/demo/repository/ConsultorioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.entity.Consultorio;
 
-public interface ConsultorioRepository extends CrudRepository<Consultorio, Integer> {
+public interface ConsultorioRepository extends CrudRepository<Consultorio, Long> {
 
 }
