+++ src/main/java/com/example/demo/repository/MedicamentoRepository.java
@@ -3,8 +3,10 @@ package com.example.demo.repository;
 import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.entity.Medicamento;
+
+
 import java.util.List;
 
-public interface MedicamentoRepository extends CrudRepository<Medicamento, Integer> {
+public interface MedicamentoRepository extends CrudRepository<Medicamento, Long> {
 	List<Medicamento> findByNombre(String nombre);
 }
